+++ passes/runtime/milp_runtime.h
@@ -38,11 +38,11 @@ extern uint16_t __nvm_sp NVM_SECTION;
 
 #ifdef DEVICE_DEBUG
 /* NVM Debug Counters */
-extern uint16_t cnt_boundary NVM_SECTION;
-extern uint16_t cnt_save_vreg NVM_SECTION;
-extern uint16_t cnt_restore_vreg NVM_SECTION;
-extern uint16_t cnt_store_mem NVM_SECTION;
-extern uint16_t cnt_restore_mem NVM_SECTION;
+extern uint32_t cnt_boundary NVM_SECTION;
+extern uint32_t cnt_save_vreg NVM_SECTION;
+extern uint32_t cnt_restore_vreg NVM_SECTION;
+extern uint32_t cnt_store_mem NVM_SECTION;
+extern uint32_t cnt_restore_mem NVM_SECTION;
 #endif
 
 /**
